+++ app/ExportTypeEnum.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kami\Cocktail;
+
+enum ExportTypeEnum: string
+{
+    case JSON = 'json';
+    case YAML = 'yml';
+}