+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-# v3.7.1
+# v3.8.0
+## New
+- Added `favorite_tags` to stats endpoint
+- Added `--type` option to recipes export
+    - Suports `yml` (default) and `json`
+
 ## Fixes
 - Optimized cocktail shelf query when parent ingredient tracking is on
+- Searching by name now also searches slug
+- Fixed missing substitute ingredient data on default recipe data import
+- Fixed missing parent ingredient data on default recipe data import
+- Fixed missing ingredient note data on default recipe data import
 
 # v3.7.0
 ## New